requirements.txt
timm==0.4.5
torchvision==0.9.1
torch==1.8.1
numpy==1.19.5
pandas==1.1.5
scipy==1.5.4
scikit-learn==0.24.1
scikit-image==0.17.2
Pillow==8.1.0
matplotlib>=2.2.2
tqdm==4.54.0
netcal==1.1.2
openpyxl>=2.5.7
